Add version.py to package_data and move back to root of repo
